+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+
+v0.02:
+* fix - --all option, remove shift
+* fix - RemoveTargetFiles(), replace break by exit
+* add - support 7zip, lz4, zip compression
+
 v0.01:
-* add - support one file, directory, all files in current directory, specific extention in directory tree, depth level choice 
+* add - support one file, directory, all files in current directory, specific extension in directory tree, depth level choice 
 * add - support tar.zx compression